refactor(app): useSidebarMenus 外移 types/constants 至 sidebarMenus/
- 2688→2365 行：types.ts 298（menu 类型+provider 选择纯函数）+ constants.ts 28
- 外部消费面经主文件 export type 锁（PinnedThreadList/Sidebar/ProviderContinuationDialog 等零改动）
- baseline 重生成 + 新文件登记豁免
验证：tsc 本刀文件零 error；useSidebarMenus.test.tsx 55 passed
注：同目录 useNormalizedRealtimePipeline.ts 为并行子代理在途，未裹入
