+++ docs/architecture/large-file-baseline.json
@@ -1,18 +1,8 @@
 {
-  "generatedAt": "2026-08-30T16:58:02.035Z",
+  "generatedAt": "2026-08-30T17:24:11.581Z",
   "scope": "fail",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
-    {
-      "path": "src-tauri/src/engine/commands.rs",
-      "lines": 5432,
-      "type": "rust",
-      "policyId": "bridge-runtime-critical",
-      "priority": "P0",
-      "warnThreshold": 2200,
-      "failThreshold": 2600,
-      "thresholdSource": "policy"
-    },
     {
       "path": "src-tauri/src/shared_session_v2.rs",
       "lines": 5349,
@@ -31,6 +21,16 @@
       "priority": "P1",
       "warnThreshold": 2600,
       "failThreshold": 3000,
+      "thresholdSource": "policy"
+    },
+    {
+      "path": "src-tauri/src/engine/commands.rs",
+      "lines": 5047,
+      "type": "rust",
+      "policyId": "bridge-runtime-critical",
+      "priority": "P0",
+      "warnThreshold": 2200,
+      "failThreshold": 2600,
       "thresholdSource": "policy"
     },
     {
